@@ -3,26 +3,25 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using MediatR;
 
-namespace RazorPagesLab.Pages.AddressBook
+namespace RazorPagesLab.Pages.AddressBook;
+
+public class CreateAddressRequest
+	: IRequest<Guid>
 {
-    public class CreateAddressRequest
-        : IRequest<Guid>
-    {
-        [Required(ErrorMessage = "Address line 1 is required.")]
-        [DisplayName("Address Line 1")]
-        public string Line1 { get; set; }
+	[Required(ErrorMessage = "Address line 1 is required.")]
+	[DisplayName("Address Line 1")]
+	public string Line1 { get; set; }
 
-        [DisplayName("Address Line 2")]
-        public string Line2 { get; set; }
+	[DisplayName("Address Line 2")]
+	public string Line2 { get; set; }
 
-        [Required(ErrorMessage = "City is required.")]
-        public string City { get; set; }
-        
-        [Required(ErrorMessage = "State is required.")]
-        public string State { get; set; }
+	[Required(ErrorMessage = "City is required.")]
+	public string City { get; set; }
 
-        [DisplayName("Postal Code")]
-        [Required(ErrorMessage = "Postal code is required.")]
-        public string PostalCode { get; set; }
-    }
+	[Required(ErrorMessage = "State is required.")]
+	public string State { get; set; }
+
+	[DisplayName("Postal Code")]
+	[Required(ErrorMessage = "Postal code is required.")]
+	public string PostalCode { get; set; }
 }
