@@ -1,20 +1,19 @@
 ﻿using System;
 using System.Linq.Expressions;
 
-namespace RazorPagesLab.Pages.AddressBook
+namespace RazorPagesLab.Pages.AddressBook;
+
+public class EntryByIdSpecification : Specification<AddressBookEntry>
 {
-	public class EntryByIdSpecification : Specification<AddressBookEntry>
+	private readonly Guid _id;
+
+	public EntryByIdSpecification(Guid id)
 	{
-		private readonly Guid _id;
+		_id = id;
+	}
 
-		public EntryByIdSpecification(Guid id)
-		{
-			_id = id;
-		}
-
-		public override Expression<Func<AddressBookEntry, bool>> ToExpression()
-		{
-			return entry => entry.Id == _id;
-		}
+	public override Expression<Func<AddressBookEntry, bool>> ToExpression()
+	{
+		return entry => entry.Id == _id;
 	}
 }
