@@ -1,21 +1,20 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace RazorPagesLab.Pages.AddressBook
+namespace RazorPagesLab.Pages.AddressBook;
+
+public class IndexModel : PageModel
 {
-    public class IndexModel : PageModel
+	private readonly IRepo<AddressBookEntry> _repo;
+	public IEnumerable<AddressBookEntry> AddressBookEntries;
+
+	public IndexModel(IRepo<AddressBookEntry> repo)
 	{
-		private readonly IRepo<AddressBookEntry> _repo;
-		public IEnumerable<AddressBookEntry> AddressBookEntries;
+		_repo = repo;
+	}
 
-		public IndexModel(IRepo<AddressBookEntry> repo)
-		{
-			_repo = repo;
-		}
-
-        public void OnGet()
-		{
-			AddressBookEntries = _repo.Find(new AllEntriesSpecification());
-		}
-    }
+	public void OnGet()
+	{
+		AddressBookEntries = _repo.Find(new AllEntriesSpecification());
+	}
 }
