@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Linq.Expressions;
 
-namespace RazorPagesLab
+namespace RazorPagesLab;
+
+public abstract class Specification<T>
 {
-	public abstract class Specification<T>
-	{
-		public abstract Expression<Func<T, bool>> ToExpression();
-	}
+	public abstract Expression<Func<T, bool>> ToExpression();
 }
