@@ -5,40 +5,39 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace RazorPagesLab
+namespace RazorPagesLab;
+
+public class Startup
 {
-	public class Startup
+	public Startup(IConfiguration configuration)
 	{
-		public Startup(IConfiguration configuration)
-		{
-			Configuration = configuration;
-		}
+		Configuration = configuration;
+	}
 
-		public IConfiguration Configuration { get; }
+	public IConfiguration Configuration { get; }
 
-		// This method gets called by the runtime. Use this method to add services to the container.
-		public void ConfigureServices(IServiceCollection services)
-		{
-			services.AddRazorPages();
-			services.AddMediatR(typeof(Startup).Assembly);
-			services.AddSingleton(typeof(IRepo<>), typeof(Repository<>));
-		}
+	// This method gets called by the runtime. Use this method to add services to the container.
+	public void ConfigureServices(IServiceCollection services)
+	{
+		services.AddRazorPages();
+		services.AddMediatR(typeof(Startup).Assembly);
+		services.AddSingleton(typeof(IRepo<>), typeof(Repository<>));
+	}
 
-		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-		{
-			if (env.IsDevelopment())
-				app.UseDeveloperExceptionPage();
-			else
-				app.UseExceptionHandler("/Error");
+	// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+	public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+	{
+		if (env.IsDevelopment())
+			app.UseDeveloperExceptionPage();
+		else
+			app.UseExceptionHandler("/Error");
 
-			app.UseStaticFiles();
+		app.UseStaticFiles();
 
-			app.UseRouting();
+		app.UseRouting();
 
-			app.UseAuthorization();
+		app.UseAuthorization();
 
-			app.UseEndpoints(endpoints => { endpoints.MapRazorPages(); });
-		}
+		app.UseEndpoints(endpoints => { endpoints.MapRazorPages(); });
 	}
 }
