@@ -2,5 +2,5 @@ using System;
 
 public interface IEntity
 {
-    Guid Id { get; }
+	Guid Id { get; }
 }
