@@ -1,48 +1,70 @@
 ﻿@page
-@model RazorPagesLab.Pages.AddressBook.EditModel
+@model EditModel
 @{
-    ViewData["Title"] = "Edit Address Book Entry";
+	ViewData["Title"] = "Edit Address Book Entry";
 }
 
 <style>
-    .validation-summary-valid {
-        display: none;
-    }
+	.validation-summary-valid { display: none; }
 </style>
 
 <h1>Address Book Entry Details</h1>
 
-<form asp-page="Edit" method="post">
-    <span>@ViewData["Message"]</span>
+<form asp-page="Edit"
+	method="post">
+	<span>@ViewData["Message"]</span>
 
-    <div asp-validation-summary="All" class="validation-summary-errors">
-        <span>Please correct the following errors:</span>
-    </div>
-    
-    <div class="form-group">
+	<div asp-validation-summary="All"
+		class="validation-summary-errors">
+		<span>Please correct the following errors:</span>
+	</div>
+
+	<div class="form-group">
 		<span asp-for="UpdateAddressRequest.Id"></span>
-        <input asp-for="UpdateAddressRequest.Id" type="hidden" />
-    </div>
-    <div class="form-group">
-        <label asp-for="UpdateAddressRequest.Line1"></label>
-        <input asp-for="UpdateAddressRequest.Line1" type="text" class="form-control" id="line1" placeholder="Line 1" />
-    </div>
-    <div class="form-group">
-        <label asp-for="UpdateAddressRequest.Line2"></label>
-        <input asp-for="UpdateAddressRequest.Line2" type="text" class="form-control" id="line2" placeholder="Line 2" />
-    </div>
-    <div class="form-group">
-        <label asp-for="UpdateAddressRequest.City"></label>
-        <input asp-for="UpdateAddressRequest.City" type="text" class="form-control" id="City" placeholder="City" />
-    </div>
-    <div class="form-group">
-        <label asp-for="UpdateAddressRequest.State"></label>
-        <input asp-for="UpdateAddressRequest.State" type="text" class="form-control" id="State" placeholder="State" />
-    </div>
-    <div class="form-group">
-        <label asp-for="UpdateAddressRequest.PostalCode"></label>
-        <input asp-for="UpdateAddressRequest.PostalCode" type="text" class="form-control" placeholder="Postal Code" />
-    </div>
-    <button type="submit" class="btn btn-primary">Update Address</button>
+		<input asp-for="UpdateAddressRequest.Id"
+				type="hidden" />
+	</div>
+	<div class="form-group">
+		<label asp-for="UpdateAddressRequest.Line1"></label>
+		<input asp-for="UpdateAddressRequest.Line1"
+				type="text"
+				class="form-control"
+				id="line1"
+				placeholder="Line 1" />
+	</div>
+	<div class="form-group">
+		<label asp-for="UpdateAddressRequest.Line2"></label>
+		<input asp-for="UpdateAddressRequest.Line2"
+				type="text"
+				class="form-control"
+				id="line2"
+				placeholder="Line 2" />
+	</div>
+	<div class="form-group">
+		<label asp-for="UpdateAddressRequest.City"></label>
+		<input asp-for="UpdateAddressRequest.City"
+				type="text"
+				class="form-control"
+				id="City"
+				placeholder="City" />
+	</div>
+	<div class="form-group">
+		<label asp-for="UpdateAddressRequest.State"></label>
+		<input asp-for="UpdateAddressRequest.State"
+				type="text"
+				class="form-control"
+				id="State"
+				placeholder="State" />
+	</div>
+	<div class="form-group">
+		<label asp-for="UpdateAddressRequest.PostalCode"></label>
+		<input asp-for="UpdateAddressRequest.PostalCode"
+				type="text"
+				class="form-control"
+				placeholder="Postal Code" />
+	</div>
+	<button type="submit"
+			class="btn btn-primary">
+		Update Address
+	</button>
 </form>
-
