@@ -1,27 +1,34 @@
 ﻿@page
-@model RazorPagesLab.Pages.AddressBook.IndexModel
+@model IndexModel
 @{
-    ViewData["Title"] = "Address Book";
+	ViewData["Title"] = "Address Book";
 }
 
 <h1>Address Book Entries</h1>
 
 <div class="container-fluid">
-    <a asp-page="Create">Create New Entry</a>
+	<a asp-page="Create">Create New Entry</a>
 
-    <ul class="list-group">
-        @foreach (var entry in Model.AddressBookEntries)
-        {
-            <li class="list-group-item">
-                <div class="row">
-                    <div class="col-md-8">
-                        @($"{entry.Line1} " +
-                                  $"{entry.City}, {entry.State} {entry.PostalCode}")
-                    </div>
-                    <div class="col-md-2"><a asp-page="Edit" asp-route-id="@entry.Id">Edit</a></div>
-                    <div class="col-md-2"><a asp-page="Delete">Delete</a></div>
-                </div>
-            </li>
-        }
-    </ul>
+	<ul class="list-group">
+		@foreach (var entry in Model.AddressBookEntries)
+		{
+			<li class="list-group-item">
+				<div class="row">
+					<div class="col-md-8">
+						@($"{entry.Line1} " +
+						$"{entry.City}, {entry.State} {entry.PostalCode}")
+					</div>
+					<div class="col-md-2">
+						<a asp-page="Edit"
+							asp-route-id="@entry.Id">
+							Edit
+						</a>
+					</div>
+					<div class="col-md-2">
+						<a asp-page="Delete">Delete</a>
+					</div>
+				</div>
+			</li>
+		}
+	</ul>
 </div>
