@@ -1,7 +1,7 @@
 ﻿@page
 @model ErrorModel
 @{
-    ViewData["Title"] = "Error";
+	ViewData["Title"] = "Error";
 }
 
 <h1 class="text-danger">Error.</h1>
@@ -9,18 +9,18 @@
 
 @if (Model.ShowRequestId)
 {
-    <p>
-        <strong>Request ID:</strong> <code>@Model.RequestId</code>
-    </p>
+	<p>
+		<strong>Request ID:</strong> <code>@Model.RequestId</code>
+	</p>
 }
 
 <h3>Development Mode</h3>
 <p>
-    Swapping to the <strong>Development</strong> environment displays detailed information about the error that occurred.
+	Swapping to the <strong>Development</strong> environment displays detailed information about the error that occurred.
 </p>
 <p>
-    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
-    It can result in displaying sensitive information from exceptions to end users.
-    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
-    and restarting the app.
+	<strong>The Development environment shouldn't be enabled for deployed applications.</strong>
+	It can result in displaying sensitive information from exceptions to end users.
+	For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
+	and restarting the app.
 </p>
